requirements.txt
DJANGO>=4.2,<5.0
djangorestframework>=3.14,<3.15
djangorestframework-simplejwt>=5.3,<6.0
python-dotenv>=1.0,<2.0
psycopg2-binary>=2.9,<3.0
django-cors-headers>=4.3,<5.0
dj-database-url>=2.1,<3.0
gunicorn>=21.2,<22.0
whitenoise>=6.7,<7.0
whitenoise
sentence-transformers>=2.0,<3.0
scikit-learn>=1.0,<2.0
numpy>=1.20,<2.0
pandas>=1.0,<3.0
graphene-django>=3.1,<4.0
